# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LlmConnectionCreate": {
        "properties": {
          "allowed_models": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Allowed Models",
            "type": "array"
          },
          "audit_mode": {
            "default": "redacted",
            "enum": [
              "full",
              "redacted",
              "off"
            ],
            "title": "Audit Mode",
            "type": "string"
          },
          "base_config": {
            "additionalProperties": true,
            "title": "Base Config",
            "type": "object"
          },
          "cache_ttl_s": {
            "default": 0,
            "maximum": 86400.0,
            "minimum": 0.0,
            "title": "Cache Ttl S",
            "type": "integer"
          },
          "data_residency": {
            "enum": [
              "local",
              "eu",
              "external"
            ],
            "title": "Data Residency",
            "type": "string"
          },
          "guardrail_policy_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardrail Policy Id"
          },
          "legal_basis": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Legal Basis",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "enum": [
              "echo",
              "openai",
              "anthropic",
              "ollama"
            ],
            "title": "Provider",
            "type": "string"
          },
          "purpose_codes": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Purpose Codes",
            "type": "array"
          },
          "secret": {
            "anyOf": [
              {
                "maxLength": 4000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret"
          }
        },
        "required": [
          "name",
          "provider",
          "data_residency",
          "legal_basis",
          "purpose_codes"
        ],
        "title": "LlmConnectionCreate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "OSAIP API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/llm-connections": {
      "get": {
        "operationId": "list_global_connections",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Global Connections"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Global Connections",
        "tags": [
          "llm-connections"
        ]
      },
      "post": {
        "operationId": "create_global_connection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LlmConnectionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Global Connection",
        "tags": [
          "llm-connections"
        ]
      }
    },
...
    "/api/v1/projects/{key}/llm-connections": {
      "get": {
        "operationId": "list_project_connections",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Project Connections"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Project Connections",
        "tags": [
          "llm-connections"
        ]
      },
      "post": {
        "operationId": "create_project_connection",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LlmConnectionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project Connection",
        "tags": [
          "llm-connections"
        ]
      }
    },
    "/api/v1/projects/{key}/llm-connections/{connection_id}": {
      "delete": {
        "operationId": "archive_project_connection",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Project Connection",
        "tags": [
          "llm-connections"
        ]
      },
      "get": {
        "operationId": "get_project_connection",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Project Connection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project Connection",
        "tags": [
          "llm-connections"
        ]
      },
      "patch": {
        "operationId": "update_project_connection",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LlmConnectionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Update Project Connection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Project Connection",
        "tags": [
          "llm-connections"
        ]
      }
    },
    "/api/v1/projects/{key}/llm-connections/{connection_id}/test": {
      "post": {
        "description": "Send one tiny call through the MESH, not around it.\n\nTesting by calling the provider directly here would prove the credentials work\nwhile bypassing exactly the layer that matters \u2014 the residency gate, the\nguardrails, the ledger. A test that skips them tests the wrong thing.",
        "operationId": "test_project_connection",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Test Project Connection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Project Connection",
        "tags": [
          "llm-connections"
        ]
      }
    },
...
  }
}